// src/api/generated/models/EmployeeActivityRate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Blackwell Time API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * id is omitted for a newly-added row that hasn't been saved yet.
 * @export
 * @interface EmployeeActivityRate
 */
export interface EmployeeActivityRate {
    /**
     * 
     * @type {string}
     * @memberof EmployeeActivityRate
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof EmployeeActivityRate
     */
    activityId: string;
    /**
     * 
     * @type {EmployeeActivityRatePayRateTypeEnum}
     * @memberof EmployeeActivityRate
     */
    payRateType: EmployeeActivityRatePayRateTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof EmployeeActivityRate
     */
    payRate: number;
    /**
     * 
     * @type {number}
     * @memberof EmployeeActivityRate
     */
    holidayPayRate: number;
}


/**
 * @export
 */
export const EmployeeActivityRatePayRateTypeEnum = {
    Hourly: 'Hourly',
    FlatRate: 'FlatRate',
    Salary: 'Salary'
} as const;
export type EmployeeActivityRatePayRateTypeEnum = typeof EmployeeActivityRatePayRateTypeEnum[keyof typeof EmployeeActivityRatePayRateTypeEnum];


/**
 * Check if a given object implements the EmployeeActivityRate interface.
 */
export function instanceOfEmployeeActivityRate(value: object): value is EmployeeActivityRate {
    if (!('activityId' in value) || value['activityId'] === undefined) return false;
    if (!('payRateType' in value) || value['payRateType'] === undefined) return false;
    if (!('payRate' in value) || value['payRate'] === undefined) return false;
    if (!('holidayPayRate' in value) || value['holidayPayRate'] === undefined) return false;
    return true;
}

export function EmployeeActivityRateFromJSON(json: any): EmployeeActivityRate {
    return EmployeeActivityRateFromJSONTyped(json, false);
}

export function EmployeeActivityRateFromJSONTyped(json: any, ignoreDiscriminator: boolean): EmployeeActivityRate {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'activityId': json['activityId'],
        'payRateType': json['payRateType'],
        'payRate': json['payRate'],
        'holidayPayRate': json['holidayPayRate'],
    };
}

export function EmployeeActivityRateToJSON(json: any): EmployeeActivityRate {
    return EmployeeActivityRateToJSONTyped(json, false);
}

export function EmployeeActivityRateToJSONTyped(value?: EmployeeActivityRate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'activityId': value['activityId'],
        'payRateType': value['payRateType'],
        'payRate': value['payRate'],
        'holidayPayRate': value['holidayPayRate'],
    };
}